Release v3.5.1: read the lunar columns as numbers, not as text
The Hijri columns are DOUBLE in the parquet, and every code path that rendered
them as text cast straight to VARCHAR. CAST(3.0 AS VARCHAR) is '3.0', and
DuckDB's lpad TRUNCATES a string already longer than the target width instead of
leaving it alone, so lpad(CAST(hijri_month AS VARCHAR), 2, '0') produced '3.'
rather than '03'. Months 10, 11 and 12 came out '10', '11', '12' by accident,
which is exactly why the fault survived: three of the twelve buckets looked
right.

get_temporal_distribution(granularity=lunar_month) therefore returned nine keys
no month_labels entry matches, and the charts.html lunar view drew three bars out
of twelve — 4,563 of the 13,261 dated articles. Nothing errored anywhere. The
chart's annotation then divided that truncated total by twelve and reported "Dhu
al-Hijja leads at 1854 (+388% vs an even split)", a figure produced entirely by
the missing bars: Ramadan actually leads at 1918, and the real spread is +74%.
The hijri month and year branches were malformed the same way, as '1420.0-3.'
and '1420.0'.

HIJRI_DATE_EXPR carried the identical bug with a wider blast radius. It builds
the hijri_date field on every article and publication row, and because the year
part is unpadded there was no accidental-pass case at all: all 13,261 values were
malformed, reading 1440.0-3.-29 instead of 1440-03-29. That is the one users saw
most and the one nothing was checking.

Both now go through hijriPart(), which carries the DOUBLE -> INTEGER cast in one
place with the reason attached, so the two call sites cannot drift apart again.
It is also robust to the columns changing type: CAST('03' AS INTEGER) is still 3.

The reason the suite never caught any of this is the fixture, not the assertions.
make-fixtures declared the three columns BIGINT while the real parquet stores
DOUBLE, and CAST(3 AS VARCHAR) is '3' — so the fixtures rendered every bucket and
every date correctly no matter how the SQL was written, and could not reproduce
this class of bug at all. Typing them DOUBLE reproduces all four failures on
unmodified source. That change is worth more than the fix it caught.

On top of it, the lunar_month check now asserts the two key sets as a contract:
month_labels is exactly 01..12, and every distribution key is present in it. A
bucket key that misses is not an error anywhere in the stack — it is a bar that
silently never draws — so it has to be asserted rather than inferred from counts.
The weekly live smoke test pins counts but never cross-checked the two sets,
which is why its pins stayed green throughout.

The README and skill figures are re-measured against the corrected output:
Ramadan +74%, Dhu al-Hijja +68%, Shawwal +42% over 13,261 fully-dated articles,
with the six ordinary months 24-32% below an even split. Rabi' I holds at -5%, so
the Maouloud finding is unchanged. The stated conversion rate moves 98.9% -> 99%
with the current dataset; that one is drift, not this bug.

+++ mcpb/scripts/make-fixtures.mjs
@@ -425,8 +425,14 @@ async function main() {
     // range) are left NULL — an imprecise date has no lunar day, and that
     // absence is a case the tools have to report rather than plot.
     if (HIJRI_SUBSETS.includes(subset)) {
+      // DOUBLE, not BIGINT, because that is what the real parquet stores — and
+      // the difference is not cosmetic. `CAST(3 AS VARCHAR)` is '3' but
+      // `CAST(3.0 AS VARCHAR)` is '3.0', so a fixture typed BIGINT renders
+      // every Hijri bucket and date correctly no matter how the SQL is written,
+      // and cannot reproduce a whole class of formatting bug that production
+      // hits on every row. It hid exactly that once already.
       for (const col of ["hijri_year", "hijri_month", "hijri_day"]) {
-        await conn.run(`ALTER TABLE ${table} ADD COLUMN "${col}" BIGINT`);
+        await conn.run(`ALTER TABLE ${table} ADD COLUMN "${col}" DOUBLE`);
       }
       for (const [greg, [hy, hm, hd]] of Object.entries(HIJRI)) {
         await conn.run(
